+++ lib/services/post.dart
@@ -11,16 +11,20 @@ import 'package:waves/constants.dart';
 class PostService {
   static final _firestore = FirebaseFirestore.instance;
   static final _storage = FirebaseStorage.instance;
-  static final List<dynamic> _posts = [];
+  static List<Post> _posts = [];
   static final List<dynamic> _images = [];
 
-  static List<dynamic> get post => _posts;
+  static Future<List<Post>> get post async {
+    await _fetchPosts();
+    _posts = sortingPost(posts: _posts);
+    return _posts;
+  }
 
   static List<dynamic> get image => _images;
   static final flutterLocalNotificationsPlugin =
-  FlutterLocalNotificationsPlugin();
+      FlutterLocalNotificationsPlugin();
 
-  static Future<void> fetchPosts() async {
+  static Future<void> _fetchPosts() async {
     await _firestore.collection('counter').doc('counter').get();
     final sourceFromDatabase = await _firestore.collection('posts').get();
     for (var post in sourceFromDatabase.docs) {
@@ -41,13 +45,11 @@ class PostService {
   }
 
   static NotiType _determineNotiType({required int rating}) {
-    if (rating <= 2 ) {
+    if (rating <= 2) {
       return NotiType.immediate;
-    }
-    else if (rating <= 3) {
+    } else if (rating <= 3) {
       return NotiType.normal;
-    }
-    else {
+    } else {
       return NotiType.fine;
     }
   }
@@ -128,5 +130,21 @@ class PostService {
         return '${duration.inSeconds} seconds later';
       }
     }
+  }
+
+  static sortingPost({required List<Post> posts}) {
+    for (int i = 0; i < posts.length; i++) {
+      for (int j = 0; j < posts.length - 1; j++) {
+        if (posts[j].lastUpdate.compareTo(posts[j + 1].lastUpdate) < 0) {
+          Post post = posts[j];
+          posts[j] = posts[j + 1];
+          posts[j + 1] = post;
+        }
+      }
+    }
+    for (Post post in posts) {
+      print(post.lastUpdate);
+    }
+    return posts;
   }
 }
